mlops_level3/ml_pipeline/pyproject.toml
[project]
name = "ml_pipeline"
requires-python = ">=3.10,<3.14"
version = "0.1.0"
dependencies = [
    "dagster==1.12.2",
    "fastapi>=0.121.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "ml_pipeline"
registry_modules = [
    "ml_pipeline.components.*",
]


